tutorials/09 - Image Captioning/requirements.txt
matplotlib==2.0.0
nltk==3.2.2
numpy==1.12.0
Pillow==4.0.0
argparse